+++ resources/gui/metadata/packageGenerator/controller.js
@@ -23,7 +23,6 @@ let branches = [];
 const vscode = acquireVsCodeApi();
 window.addEventListener('message', event => {
     let eventData = event.data;
-    console.log('eventData\n' + JSON.stringify(event.data, null, 2));
     switch (eventData.command) {
         case 'open':
             // @ts-ignore
@@ -105,14 +104,12 @@ window.addEventListener('message', event => {
 function loadMetadata(loadFrom) {
     // @ts-ignore
     closeAllPageMessages();
-    console.log('@@Load Metadata');
     // @ts-ignore
     openSpinnerModal();
     vscode.postMessage({ command: "loadMetadata", loadFrom: loadFrom, selectedOptionToDownload: selectedOptionToDownload });
 }
 
 function drawMetadataTypes(deleteChilds) {
-    console.log("Draw Metadata Types");
     let content = [];
     if (metadata) {
         content.push('<ul class="metadataList">');
@@ -143,7 +140,6 @@ function drawMetadataTypes(deleteChilds) {
 }
 
 function drawMetadataObjects(typeName, objects, deleteChilds) {
-    console.log("Draw Metadata Objects");
     let content = [];
     if (objects) {
         content.push('<ul class="metadataList">');
@@ -170,7 +166,6 @@ function drawMetadataObjects(typeName, objects, deleteChilds) {
 }
 
 function drawMetadataItems(typeName, objectName, items) {
-    console.log("Draw Metadata Items");
     let content = [];
     if (items) {
         content.push('<ul class="metadataList">');
@@ -194,7 +189,6 @@ function drawMetadataItems(typeName, objectName, items) {
 }
 
 function clickOnMetadataType(typeName) {
-    console.log('Click on Metadata Type: ' + typeName);
     nClicksOnObject = 0;
     lastObjectClicked = undefined;
     if (lastTypeClicked === typeName)
@@ -205,7 +199,6 @@ function clickOnMetadataType(typeName) {
     if (metadata) {
         let metadataType = metadata[typeName];
         if (metadataType) {
-            console.log('metadataType: ' + metadataType.name);
             if (Object.keys(metadataType.childs).length > 0) {
                 if (nClicksOnMetadata > 0) {
                     metadataType.checked = !metadataType.checked;
@@ -236,7 +229,6 @@ function clickOnMetadataType(typeName) {
 }
 
 function clickOnMetadataObject(typeName, objName) {
-    console.log('Click on Metadata Object: ' + typeName + '.' + objName);
     nClicksOnMetadata = 0;
     lastTypeClicked = undefined;
     if (lastObjectClicked === objName)
@@ -246,10 +238,8 @@ function clickOnMetadataObject(typeName, objName) {
     selectedMetadataObject = objName;
     if (metadata) {
         let metadataType = metadata[typeName];
-        console.log('metadataType: ' + metadataType.name);
         if (metadataType) {
             let metadataObject = metadataType.childs[objName];
-            console.log('metadataObject: ' + metadataObject.name);
             if (metadataObject) {
                 if (Object.keys(metadataObject.childs).length > 0) {
                     if (nClicksOnObject > 0) {
@@ -630,7 +620,6 @@ function selectAll() {
     let component = document.getElementById('selectAll');
     let selectAll = (component.innerHTML === 'All') ? true : false;
     let linkText = (component.innerHTML === 'All') ? 'None' : 'All';
-    console.log(component.innerHTML);
     if (metadata) {
         if (selectAll) {
             checkAll(metadata);
@@ -646,7 +635,6 @@ function selectFromPackage(pkg) {
         vscode.postMessage({ command: "selectFromPackage" });
     }
     else if (metadata) {
-        console.log(JSON.stringify(pkg, null, 2));
         Object.keys(pkg).forEach(function (type) {
             if (metadata[type]) {
                 if (pkg[type].includes('*')) {
@@ -820,14 +808,11 @@ function changeMetadataForSelect(inputId) {
 }
 
 function onChangeSaveOn(elementId) {
-    console.log(elementId);
     let element = document.getElementById(elementId);
-    console.log(element);
     // @ts-ignore
     if (element.checked)
         // @ts-ignore
         saveOn = element.value;
-    console.log(saveOn);
 }
 
 function onClickChangeFromGit(inputId) {