+++ resources/gui/metadata/permissionSet/controller.js
@@ -550,10 +550,6 @@ function getTabVisibilitySelect(tab, selectedValue) {
 }
 
 function clickOnElementCheck(section, elementName, elementCheckField) {
-    console.log('clickOnElementCheck');
-    console.log('section => ' + section);
-    console.log('elementName => ' + elementName);
-    console.log('elementCheckField => ' + elementCheckField);
     if (section === 'applicationVisibilities') {
         // @ts-ignore
         if (elementCheckField === 'default' && document.getElementById('app_' + elementName + '_' + elementCheckField).checked) {