+++ resources/scripts/commands/orgCompare.js
@@ -1,10 +1,10 @@
 const vscode = require('vscode');
 const fileSystem = require('../fileSystem');
 const ProcessManager = require('../processes').ProcessManager;
-const ProcessEvent = require('../processes').ProcessEvent;
 const Config = require('../core/config');
 const Metadata = require('../metadata');
 const GUIEngine = require('../guiEngine');
+const MetadataSelectorInput = require('../inputs/metadataSelector');
 const Engine = GUIEngine.Engine;
 const Routing = GUIEngine.Routing;
 const Paths = fileSystem.Paths;
@@ -18,44 +18,69 @@ exports.run = function () {
         title: "Comparing Org Metadata with your local metadata \n (Only affects metadata types that you have in your local project)",
         cancellable: true
     }, (progress, cancelToken) => {
-        return new Promise(async function (promiseResolve) {
+        return new Promise(async function (resolve) {
             ProcessManager.auraHelperOrgCompare(true, cancelToken).then(function (out) {
                 if (!out) {
                     vscode.window.showWarningMessage('Operation cancelled by user');
                 } else if (out.stdOut) {
                     let response = JSON.parse(out.stdOut);
                     if (response.status === 0) {
-                        let metadata = response.result.data;
-                        let viewOptions = Engine.getViewOptions();
-                        viewOptions.title = 'Match Org Metadata with Local';
-                        viewOptions.showActionBar = true;
-                        viewOptions.actions.push(Engine.createButtonAction('deleteBtn', '{!label.delete}', ["w3-btn save"], "deleteMetadata()"));
-                        viewOptions.actions.push(Engine.createButtonAction('cancelBtn', '{!label.cancel}', ["w3-btn cancel"], "cancel()"));
-                        view = Engine.createView(Routing.MatchOrg, viewOptions);
-                        view.render(metadata);
-                        view.onReceiveMessage(function (message) {
-                            if (message.command === 'cancel')
-                                view.close();
-                            else if (message.command === 'delete')
-                                deleteMetadata(message.model);
-                        });
+                        if (Config.getConfig().graphicUserInterface.enableAdvanceGUI) {
+                            openAdvanceGUI(response.result.data);
+                        } else {
+                            openStandardGUI(response.result.data);
+                        }
                     } else {
                         vscode.window.showErrorMessage(response.error.message);
                     }
                 } else {
                     vscode.window.showErrorMessage('An error ocurred while processing command. Error: \n' + out.stdErr);
                 }
-                promiseResolve();
+                resolve();
             }).catch(function (error) {
                 vscode.window.showErrorMessage('An error ocurred while processing command. Error: \n' + error);
-                promiseResolve();
+                resolve();
             });
         });
     });
 }
 
-async function deleteMetadata(metadataToMatch) {
-    let metadataOnOrg = metadataToMatch.metadataOnOrg;
+function openAdvanceGUI(metadata) {
+    let viewOptions = Engine.getViewOptions();
+    viewOptions.title = '{!label.match_metadata_with_local}';
+    viewOptions.showActionBar = true;
+    viewOptions.actions.push(Engine.createButtonAction('deleteBtn', '{!label.delete}', ["w3-btn save"], "deleteMetadata()"));
+    viewOptions.actions.push(Engine.createButtonAction('cancelBtn', '{!label.cancel}', ["w3-btn cancel"], "cancel()"));
+    view = Engine.createView(Routing.MatchOrg, viewOptions);
+    view.render(metadata);
+    view.onReceiveMessage(function (message) {
+        if (message.command === 'cancel')
+            view.close();
+        else if (message.command === 'delete')
+            deleteMetadata(message.model.metadataOnOrg);
+    });
+}
+
+function openStandardGUI(metadata) {
+    let input = new MetadataSelectorInput('Select Metadata Types for Delete from your Org', metadata);
+    input.allowDelete(true);
+    input.onDelete(async metadata => {
+        deleteMetadata(metadata, function (message, isError) {
+            if (isError) {
+                vscode.window.showErrorMessage(message);
+            } else {
+                metadata  = Metadata.Utils.deleteCheckedMetadata(metadata);
+                vscode.window.showInformationMessage(message);
+                input.setMetadata(metadata);
+                input.reset();
+            }
+        });
+    });
+    input.show();
+}
+
+async function deleteMetadata(metadataToMatch, callback) {
+    let metadataOnOrg = metadataToMatch;
     let version = Config.getOrgVersion();
     let user = await Config.getAuthUsername();
     let folder = Paths.getPackageFolder();
@@ -72,53 +97,61 @@ async function deleteMetadata(metadataToMatch) {
         title: "Deleting Selected Metadata from Org",
         cancellable: true
     }, (progress, cancelToken) => {
-        return new Promise(promiseResolve => {
+        return new Promise(async resolve => {
             try {
-                let buffer = [];
-                let bufferError = [];
-                ProcessManager.destructiveChanges(user, folder, cancelToken, function (event, data) {
-                    switch (event) {
-                        case ProcessEvent.ERR_OUT:
-                        case ProcessEvent.ERROR:
-                            bufferError = bufferError.concat(data);
-                            break;
-                        case ProcessEvent.END:
-                            if (bufferError.length > 0)
-                                view.postMessage({ command: 'metadataDeletedError', data: { error: bufferError.toString() } });
-                            else
-                                processResponse(buffer.toString(), promiseResolve);
-                            break;
-                        case ProcessEvent.KILLED:
-                            view.postMessage({ command: 'processKilled' });
-                            break;
-                        case ProcessEvent.STD_OUT:
-                            buffer = buffer.concat(data);
-                            break;
-                        default:
-                            break;
+                let out = await ProcessManager.destructiveChanges(user, folder, cancelToken);
+                if (out) {
+                    if (out.stdOut) {
+                        let response = JSON.parse(out.stdOut);
+                        if (response.status === 0) {
+                            if (callback) {
+                                callback.call(this, 'Metadata Deleted Successfully', false);
+                            } else {
+                                view.postMessage({ command: 'metadataDeleted', data: { metadata: Metadata.Utils.deleteCheckedMetadata(metadataToMatch) } });
+                            }
+                        } else {
+                            if (callback) {
+                                let errors = '';
+                                if (response.result.details && response.result.details.componentFailures && response.result.details.componentFailures.length > 0) {
+                                    for (const fail of response.result.details.componentFailures) {
+                                        errors += '- ' + fail.fullName + ': ' + fail.problem + '\n';
+                                    }
+                                }
+                                callback.call(this, errors, true)
+                            } else {
+                                let errors = '';
+                                if (response.result.details && response.result.details.componentFailures && response.result.details.componentFailures.length > 0) {
+                                    for (const fail of response.result.details.componentFailures) {
+                                        errors += '<b>' + fail.fullName + '</b>: ' + fail.problem + '\n';
+                                    }
+                                }
+                                view.postMessage({ command: 'metadataDeletedError', data: { error: errors } });
+                            }
+                        }
+                    } else {
+                        if (callback) {
+                            callback.call(this, out.stdErr, true)
+                        } else {
+                            view.postMessage({ command: 'metadataDeletedError', data: { error: out.stdErr } });
+                        }
                     }
-                });
+                } else {
+                    if (callback) {
+                        callback.call(this, 'Operation Cancelled by User', true)
+                    } else {
+                        view.postMessage({ command: 'processKilled' });
+                    }
+                }
+                resolve();
             } catch (error) {
-                view.postMessage({ command: 'metadataDeletedError', data: { error: error } });
+                if (callback) {
+                    callback.call(this, error.toString(), true);
+                } else {
+                    view.postMessage({ command: 'metadataDeletedError', data: { error: error } });
+                }
             }
         });
     });
-}
-
-function processResponse(stdOut, promiseResolve) {
-    let jsonOut = JSON.parse(stdOut);
-    if (jsonOut.status === 0) {
-        view.postMessage({ command: 'metadataDeleted', data: {} });
-    } else if (jsonOut.status === 1) {
-        let errors = '';
-        if (jsonOut.result.details && jsonOut.result.details.componentFailures && jsonOut.result.details.componentFailures.length > 0) {
-            for (const fail of jsonOut.result.details.componentFailures) {
-                errors += '<b>' + fail.fullName + '</b>: ' + fail.problem + '\n';
-            }
-        }
-        view.postMessage({ command: 'metadataDeletedError', data: { error: errors } });
-    }
-    promiseResolve();
 }
 
 function createPackage(version, filePath, isDestructive) {